+++ src/ImageSharp/Processing/Processors/Transforms/Linear/AffineTransformProcessor.cs
