+++ src/ImageSharp/Processing/Processors/Transforms/Linear/AutoOrientProcessor.cs
