+++ src/ImageSharp/Processing/Processors/Transforms/Linear/RotateProcessor.cs
