+++ src/ImageSharp/Processing/Processors/Transforms/Linear/RotateProcessor{TPixel}.cs
